Created displayelemt function and dynamic todo data
