+++ src/template.html
@@ -22,7 +22,7 @@
           name="todo-text"
           placeholder="Add your todo list..."
         />
-        <p>↲</p>
+        <div class="add-todo-btn">↲</div>
       </div>
       <div class="todos">
         <!-- Todo goes here -->
