+++ src/utilities/DocumentEvents.js
@@ -1,5 +1,9 @@
-export const DocumentEvent = (eventType, event) => {
+const DocumentEvent = (eventType, event) => {
   document.addEventListener(eventType, event);
+};
+
+export const elEvent = (el,eventType, event) => {
+  el.addEventListener(eventType, event);
 };
 
 export default DocumentEvent;
