+++ src/utilities/todoData.js
@@ -1,10 +1,10 @@
-const Data = [
-  { index: 1, discription: 'Get enough money', completed: false },
-  { index: 2, discription: 'Get enough money', completed: false },
-  { index: 3, discription: 'Get enough money', completed: true },
-  { index: 4, discription: 'Get enough money', completed: false },
-  { index: 5, discription: 'Get enough money', completed: false },
-  { index: 6, discription: 'Get enough money', completed: true },
-];
+import { getItemsFromStorage } from "./Storage";
+let todos;
+if(!getItemsFromStorage('todos')){
+  todos = []
+}else {
+  todos = getItemsFromStorage('todos');
+}
 
-export default Data;
+
+export default todos
